# server/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Crypto Management API",
    "description": "This API provides basic crypto management functionalities using CoinGecko API. You can get real-time prices, list all cryptocurrencies, and retrieve company information for a specific coin.",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/api",
  "schemes": [
    "http"
  ],
  "paths": {
    "/GetCoins": {
      "get": {
        "description": "This will fetch all the coins since due very large number only 100 coins are listed",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/GetCoinPrice": {
      "post": {
        "description": "This will give you the  price of a coin in the given currency.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "coinName": {
                  "example": "bitcoin"
                },
                "toCurrency": {
                  "example": "eth"
                },
                "date": {
                  "example": "12-01-2023"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/company_list": {
      "post": {
        "description": "This will list all the company associated with the specific coin",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "currency": {
                  "example": "bitcoin"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  }
}